requirements.txt
Flask
python-dotenv
requests
pandas
numpy
scikit-learn
torch
tensorflow
shap
joblib
